+++ packages/frontend/src/design-system/components/common/index.ts
@@ -1,3 +1,4 @@
 export { default as Button } from "./Button";
 export { Badge, BadgeGroup } from "./Badge";
 export { toast, ToastContainer } from "./Toast";
+export { default as Accordion } from "./Accordion/Accordion";
